evaluation/experiment-1/python_yamcs_api/requirements.txt
yamcs-client==1.12.0